+++ sat/system/processor.py
@@ -5,12 +5,12 @@ Copyright 2019 Cray Inc. All Rights Reserved.
 """
 
 from sat.cached_property import cached_property
-from sat.system.component import BaseComponent
+from sat.system.component import NodeComponent
 from sat.system.constants import PROCESSOR_TYPE
 from sat.system.field import ComponentField
 
 
-class Processor(BaseComponent):
+class Processor(NodeComponent):
     """Represents a processor in the system."""
 
     hsm_type = PROCESSOR_TYPE
@@ -19,7 +19,7 @@ class Processor(BaseComponent):
 
     fields = [
         ComponentField('xname'),
-        # Allow summarization by manufacturer and model unlike in BaseComponent
+        # Allow summary by manufacturer and model unlike in BaseComponent
         ComponentField('Manufacturer', summarizable=True),
         ComponentField('Model', summarizable=True),
         ComponentField('Part Number'),